some style fixes, cleanup, and better loading for cur user stuff
