+++ src/api/constants.ts
@@ -23,6 +23,7 @@ export const CHANGE_USER_NAME = PROXY + 'https://v1.peachapi.com/stream/name';
 export const ADD_FRIEND = (username: string) =>
 	`${PROXY}https://v1.peachapi.com/stream/n/${username}/connection`;
 export const CHANGE_BIO = `${PROXY}https://v1.peachapi.com/stream/bio`;
+export const UPLOAD_IMAGE = `https://api.imgur.com/3/image`;
 
 enum ACTIONS {
 	login,
