+++ src/pages/Activity/index.tsx
@@ -26,7 +26,7 @@ export const ActivityPage = () => {
 				}
 			);
 		} catch (err) {
-			console.log(err);
+			console.error(err);
 		}
 	}, [jwt]);
